requirements.txt
Flask
bcrypt
flask_wtf
flask-csp
flask_sqlalchemy
flask-login
Flask-Mail
python-dotenv